colorplash/requirements.txt
flask==2.3.3
werkzeug==2.3.7
torch>=1.13.0
torchvision>=0.14.0
Pillow>=9.3.0
numpy>=1.23.0
fastai==1.0.61
opencv-python>=4.6.0
tqdm>=4.64.0 